엔티티를 라이브러리 모듈로 이동한다
