+++ apps/backoffice-api/src/main.ts
@@ -1,12 +1,12 @@
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app.module';
+import { BackofficeModule } from './backoffice.module';
 import { Logger } from '@nestjs/common';
 import * as Sentry from '@sentry/node';
 import { WebhookInterceptor } from '@app/common/webhook.interceptor';
 import { SentryInterceptor } from '@app/common/sentry.interceptor';
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
+  const app = await NestFactory.create(BackofficeModule);
   app.setGlobalPrefix('api/v1');
 
   Sentry.init({
