--- apps/mymovester-api/src/backoffice-api.controller.ts
@@ -1,12 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import { BackofficeApiService } from './backoffice-api.service';
-
-@Controller()
-export class BackofficeApiController {
-  constructor(private readonly backofficeApiService: BackofficeApiService) {}
-
-  @Get()
-  getHello(): string {
-    return this.backofficeApiService.getHello();
-  }
-}
