--- apps/mymovester-api/src/backoffice-api.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { BackofficeApiController } from './backoffice-api.controller';
-import { BackofficeApiService } from './backoffice-api.service';
-
-@Module({
-  imports: [],
-  controllers: [BackofficeApiController],
-  providers: [BackofficeApiService],
-})
-export class BackofficeApiModule {}
