--- apps/mymovester-api/src/backoffice-api.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@nestjs/common';
-
-@Injectable()
-export class BackofficeApiService {
-  getHello(): string {
-    return 'Hello World!';
-  }
-}
