+++ apps/mymovester-api/src/main.ts
@@ -1,12 +1,12 @@
 import { NestFactory } from '@nestjs/core';
-import { BackofficeApiModule } from './backoffice-api.module';
 import { Logger } from '@nestjs/common';
 import * as Sentry from '@sentry/node';
 import { WebhookInterceptor } from '@app/common/webhook.interceptor';
 import { SentryInterceptor } from '@app/common/sentry.interceptor';
+import { MymovesterModule } from './mymovester.module';
 
 async function bootstrap() {
-  const app = await NestFactory.create(BackofficeApiModule);
+  const app = await NestFactory.create(MymovesterModule);
 
   Sentry.init({
     dsn: process.env.SENTRY_DSN_MYMOVESTER,
