+++ apps/mymovester-api/src/mymovester.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { UserModule } from './user/user.module';
+
+@Module({
+  imports: [UserModule],
+})
+export class MymovesterModule {}
