+++ apps/mymovester-api/src/user/user.controller.ts
@@ -1,0 +1,12 @@
+import { Controller, Get } from '@nestjs/common';
+import { UserService } from './user.service';
+
+@Controller('user')
+export class UserController {
+  constructor(private userService: UserService) {}
+
+  @Get('/')
+  getUser(): string {
+    return this.userService.getUser();
+  }
+}
