+++ apps/mymovester-api/src/user/user.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class UserService {
+  getUser(): string {
+    return 'hello';
+  }
+}
