+++ libs/persistence/src/domain/base-entity.entity.ts
